+++ src/java/dao/TeacherDAOInterface.java
@@ -5,13 +5,13 @@ import java.util.List;
 
 public interface TeacherDAOInterface {
 
-    public Teacher findById(Long id);
+    public Teacher findById(int id);
 
     public void addTeacher(Teacher t);
 
     public void editTeacher(Teacher t);
 
-    public void removeTeacher(Long id);
+    public void removeTeacher(int id);
 
     public List<Teacher> getAllTeacher();
 }
